% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_Err_check.R
\name{check_err_SVI}
\alias{check_err_SVI}
\title{check_err_SVI, check_err_EX, check_err_MOD, check_err_MOD2, check_err_Stat}
\usage{
check_err_SVI(er, mess, Main_GUI)
}
\arguments{
\item{er}{result of a "try" construct, so result of the function, or error of class "try-error"}

\item{mess}{specifier of a generic "mess" widget
mess_lab specifier of a generic "mess_lab" widget (Not always needed)}

\item{Main_GUI}{specifier of the MAIN GUI widget}
}
\value{
NOTHING, but gives back control to the MAIN GUI of FRG Tool
}
\description{
Accessory functions for error handling
}
\details{
These are helper functions used to handle errors during FRG Tool execution excploiting the try-error method. \cr
If an error occurs in one of the "main" modules, the appropriate check_err function kicks in. It issues a warning message describing the kind of error, \cr
performs some clean-up (e.g., closing no longer needed GUI messages), and gives back control to the MAIN GUI.
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Nov 23, 2012
}

