% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spatype.R
\name{check_spatype}
\alias{check_spatype}
\title{check_spatype}
\usage{
check_spatype(object)
}
\arguments{
\item{object}{either an ]SP object or a character vector pointing to a vector or raster layer}
}
\value{
character vector equal to spobject, rastobject, vectfile or rastfile
}
\description{
Wrapper to check if an object passed to the function corresponds to wither a Spatial Object, 
an R raster object, a file corresponding to a vector, or a file corresponding to a raster
}

