% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_Create_Shape.R
\name{FRG_Create_Shape}
\alias{FRG_Create_Shape}
\title{FRG_Create_Shape}
\usage{
FRG_Create_Shape()
}
\value{
Shape_File : Name of the created shapefile#'
}
\description{
Function used to create the input shapefile for the FRG processing, starting from Oracle database tables
}
\details{
This function is used to call the external function used to create the input shapefile for the FRG processing, starting from Oracle database tables\cr
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Nov 8, 2012
}
