% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_def_files.R
\name{frg_def_files}
\alias{frg_def_files}
\title{frg_def_files}
\usage{
frg_def_files(opts)
}
\arguments{
\item{opts}{}

\item{out_dir}{}

\item{Start_Year}{}

\item{End_Year}{}

\item{mod_dir}{}

\item{opts$orig_shapefile}{}
}
\value{
`opts`` list contains all main file names and folders
}
