% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_download_modis.R
\name{frg_download_modis}
\alias{frg_download_modis}
\title{frg_download_modis}
\usage{
frg_download_modis(opts, yy)
}
\arguments{
\item{opts}{\code{list} of options passed from \code{frg_fullprocessing()}}

\item{yy}{Year for which images are to be downloaded and processed (NOTE:
within \code{frgtool}, different years are processed in a for cycle.)}
}
\value{
The function is called for its side effects
}
\description{
Function used to download, mosaic and reproject MODIS data, using
`MODIStsp`` functionalities
}
\details{
The function leverages
functionality form \code{MODIStsp} package (https://github.com/lbusett/MODIStsp/)
\strong{Download and Mosaicing}:
Images covering Europe for DOYS 209 and 225 of each selected year are downloaded,
mosaiced and reprojected in LAEA projection, with 250m pixels size
Resulting images are saved in TIFF format in the Originals subfolder of the
output folder selected by the user to store MODIS imafes. Separate subfolders are used
to store images related to  NDVI, QA (Quality Assurance) and Pixel Reliability.
Original HDF files are stored in the "hdf" folder, while METADATA files to
access the full time series in \code{ENVI} and \code{R} are stored in the \code{time_series}
subfolder
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
