% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_extr_stats_r.R
\name{frg_extr_roi_stats_r}
\alias{frg_extr_roi_stats_r}
\title{frg_extr_roi_stats_r}
\usage{
frg_extr_roi_stats_r(burned_shape = burned_shape, Erode_File = Erode_File,
  sVI_Folder = sVI_Folder, CLC_File_00 = CLC_File_00,
  ENV_Zones_File = ENV_Zones_File, Out_File = Out_File,
  no_data_in = no_data_in, Start_Year = Start_Year, End_Year = End_Year)
}
\arguments{
\item{Erode_File}{}

\item{CLC_File_00}{}

\item{ENV_Zones_File}{}

\item{Out_File}{}

\item{no_data_in}{}

\item{Start_Year}{}

\item{End_Year}{}

\item{ROI_File}{}

\item{sVI_File}{}
}
\description{
function to extract time series data on burnt areas from multitemporal
  SVI files
}

