% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_fastzonal.R
\name{frg_fastzonal}
\alias{frg_fastzonal}
\title{frg_fastzonal}
\usage{
frg_fastzonal(in_rts, zone_object, mask_object = NULL, start_date = NULL,
  end_date = NULL, start_band = NULL, end_band = NULL, id_field = NULL,
  FUN = "mean", out_format = "xts", small = TRUE,
  small_method = "centroids", na.rm = TRUE, verbose = FALSE,
  maxchunk = 4e+07)
}
\arguments{
\item{mask_object}{}

\item{maxchunk}{}
}
\value{
Time series
}
\description{
aaaa
}
\examples{

}

