% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_proc_gui.R
\name{FRG_Full_Proc_GUI}
\alias{FRG_Full_Proc_GUI}
\title{FRG_Full_Proc_GUI}
\usage{
FRG_Full_Proc_GUI()
}
\value{
If the user chooses 'Start' and all processing parameters are OK, the selected parameters are saved in the 'FRG_Full_Previous.RData' file
}
\description{
Function used to create the GUI for selecting the parameters for full FRG procesing
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Oct 26, 2012
}
