% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_MOD_Comp_MeanInd.R
\name{FRG_MOD_Comp_MeanInd}
\alias{FRG_MOD_Comp_MeanInd}
\title{FRG_MOD_Comp_RDVI}
\usage{
FRG_MOD_Comp_MeanInd(OutOrig_Path = OutOrig_Path, ReProc = ReProc,
  yy = yy, UI_check = UI_check, max_UI = max_UI, nodata_in = nodata_in)
}
\arguments{
\item{OutOrig_Path}{string  Folder where the original MODIS images are being processed (Selected in FRG_MOD_Sel)}

\item{ReProc}{logical If TRUE, UI is recalculated even if appropriate files are already found in the file system.}

\item{yy}{numeric Year of the processing}

\item{max_UI}{Max UI value kept when averaging MODIS data. default = 5}

\item{UI_Check}{flag If 1 , then check on usefulness index is performed when averaging MODIS data. default = 1}

\item{no_data_in}{numeric NODATA value of input files}
}
\value{
string 'DONE' if processing completed correctly
}
\description{
Accessory function used to compute  the average DVI, RDVI and NDVI maps, starting from DOY 209 and 225  data of each year and accounting for NODATA and bad quality data
}
\details{
This function is used simply to call the appropriate IDL function used to compute average VI values, starting from DOY 209 and 225 data 
 and accounting for NODATA and bad quality data issues. See the IDL function "FRG_MEAN_Indexes.pro" documentation for further info.
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Nov 23, 2012
}

