% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_MOD_Comp_RDVI.R
\name{FRG_MOD_Comp_RDVI}
\alias{FRG_MOD_Comp_RDVI}
\title{FRG_MOD_Comp_RDVI}
\usage{
FRG_MOD_Comp_RDVI(OutOrig_Path = OutOrig_Path, ReProc = ReProc, yy = yy)
}
\arguments{
\item{OutOrig_Path}{string  Folder where the original MODIS images are being processed (Selected in FRG_MOD_Sel)}

\item{ReProc}{logical If TRUE, UI is recalculated even if appropriate files are already found in the file system.}

\item{yy}{numeric Year of the processing}

\item{mess_lab}{Identifier of the gWidget usd for processing status reporting}
}
\value{
string 'DONE' if processing completed correctly
}
\description{
Accessory function used to compute RDVI from RED and NIR images exploiting IDL external function
}
\details{
This function is used simply to call the appropriate IDL function used to compute RDVI values from MODIS RED and NIR bands data
values. See documentation of "FRG_COMPUTE_RDVI.pro" IDL script for additional details
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Nov 23, 2012
}

