% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_MOD_Comp_SVI.R
\name{FRG_MOD_Comp_SVI}
\alias{FRG_MOD_Comp_SVI}
\title{FRG_MOD_Comp_SVI}
\usage{
FRG_MOD_Comp_SVI(MOD_Dir = MOD_Dir, Shape_File = Shape_File,
  CLC_File_00 = CLC_File_00, Out_Folder = Out_Folder,
  Start_Year = Start_Year, End_Year = End_Year, NKer = NKer,
  Method = Method, SRDVI = SRDVI, SNDVI = SNDVI,
  nodata_out = nodata_out, ReProcIm = ReProcIm,
  Intermediate_Folder = Intermediate_Folder)
}
\arguments{
\item{MOD_Dir}{string Main Folder containing MODIS data (i.e., the one above the "Originals" folder)}

\item{Shape_File}{string ENVI ROI file containing info on burnt areas (derived from the EFFIS shapefile)}

\item{CLC_File_00}{string ENVI file containing the CORINE land Cover map 2000, recoded to the EFFIS legend}

\item{Out_Folder}{string Main Output folder for the yearly Scaled Indexes images}

\item{Start_Year}{numeric Starting year for the analysis}

\item{End_Year}{numeric Ending Year for the analysis}

\item{NKer}{numeric width (in Km) of the moving window used for computation of scaled indexes}

\item{Method}{numeric. 1 =  0-100 ReScaling; 2 = Percentage Median difference computation \cr}

\item{SRDVI}{numeric. 1 = compute SDVI; 2 =  Don't Compute SDVI}

\item{SNDVI}{numeric. 1 =  compute SNDVI; 2 = Don't Compute SNDVI}

\item{nodata_out}{numeric nodata value of input data}

\item{ReProcIm}{numeric 1 =  Reprocess already available images; 2 = Don't Reprocess already available images}

\item{Intermediate_Folder}{string Folder where ROI files and Masks files will be placed}
}
\value{
'DONE' if all went OK, otherwise error message to be treated by "try.catch"
}
\description{
Function used to call the IDL functiosn used for computation of SRDVI and SNDVI yearly images ( FRG_Compute_MedScaled_VI.pro)
}
\details{
This function is used to call the IDL function used for computation of SDVI and SNDVI yearly images.\cr
Also calls the IDL routines needed to create the ENVI burnt areas ROI, and the ENVI burnt areas MASKS (both total and eroded!) \cr
 Yearly SRDVI (or SNDVI) images are computed starting from the corresponding Yearly Average mosaics. Computation of the scaled indexes for each pixel is done as follows:\cr
\itemize{
\item The pixels belonging to the same class of the target included in a window of width specified by the user centred on the target are identified on the basis of the CLC00 recoded map
\item Pixels showing NODATA values in the input VI image, or corresponding to areas classified as "burned" on the input ROI files are removed from the above selection
\Item Remaining pixels (i.e., same class, unburned and with good data) are used to compute the MOMENTS of the distribution of VI values of the target CLC class in the window centred on the target pixel
\Item Computed MOMENTS are use to determine an ESTIMATE of the 5th, 50th (median) 95th percentile of the distribution, using the Cornish-Fisher fourth-order Expansion method (http://www.nematrian.com/R.aspx?p=CornishFisherDerivation))
\Item Scaled indexes for the target pixel is computed exploiting this method:
	  'Percentage Deviation to Kernel Median': For each pixel, the algorithm identifies all pixels of the image corresponding to the same CLC class contained in a window of width NKer
			It then computes the median of the distribution of values of the analyzed Vegetation Index in the selecteds pixels. Scaled Index for the pixel are then computed as:
			sVI = 100 * (VI_pix - VI_median)/(VI_median).
}#' \cr
(For further details, see the documentation for the IDL functions (FRG_Compute_MedScaled_VI.pro))
}
\author{
Lorenzo Busetto (2013)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Nov 8, 2013
}

