% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_MOD_Comp_UI.R
\name{FRG_MOD_Comp_UI}
\alias{FRG_MOD_Comp_UI}
\title{FRG_MOD_Comp_UI}
\usage{
FRG_MOD_Comp_UI(OutOrig_Path = OutOrig_Path, ReProc = ReProc, yy = yy)
}
\arguments{
\item{OutOrig_Path}{string  Folder where the original MODIS images are being processed (Selected in FRG_MOD_Sel)}

\item{ReProc}{logical If TRUE, UI is recalculated even if appropriate files are already found in the file system.}

\item{yy}{numeric Year of the processing}

\item{mess_lab}{Identifier of the gWidget usd for processing status reporting}
}
\value{
string 'DONE' if processing completed correctly
}
\description{
Accessory function used to perform Convertion of QA to UI values  exploiting IDL external function
}
\details{
This function is used simply to call the appropraite IDL function used to convert MODIS QA data to UI (Usefulness Index)
values. See documentation of "FRG_CONVERT_QA.pro" IDL script for additional details
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Nov 23, 2012
}

