% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_MOD_Download.R
\name{FRG_MOD_Download}
\alias{FRG_MOD_Download}
\title{FRG_MOD_Proc}
\usage{
FRG_MOD_Download(dates = dates, OutOrig_Path = OutOrig_Path,
  ReDown = ReDown, yy = yy)
}
\arguments{
\item{dates}{string array of dates to be downloaded}

\item{OutOrig_Path}{string Main folder where the original MODIS mosaics (i.e., one pan european image for each data and data type) will be stored}

\item{ReDown}{numeric if = 1, MODIS images needed to create already existing mosaic files will be redownloaded, and
already existing mosaics will be overwritten}

\item{yy}{numeric Year considered}
}
\value{

}
\description{
Function used to download, mosaic and reproject MODIS data
}
\details{
Function used to download, mosaic and reproject MODIS data  \cr
  \itemize{
		\item Download and Mosaicing: \cr
              Images covering Europe for DOYS 209 and 225 of each selected year are downloaded, mosaiced and reprojected \cr
              Resulting images are saved in TIFF format in the Originals subfolder of the output folder selected by the user. Separate subfolders are used\cr
              to store images related to  NDVI, RED and NIR reflectance, QA (Quality Assurance) and Pixel Reliability.}
}
\author{
Babak Naimi (naimi@r-gis.net)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Oct 29, 2012
}

