% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_MOD_GUI.R
\name{FRG_MOD_GUI}
\alias{FRG_MOD_GUI}
\title{FRG_MOD_GUI}
\usage{
FRG_MOD_GUI()
}
\value{
If the user chooses 'Start' and all processing parameters are OK, the selected parameters are saved in the 'FRG_MOD_Previous.RData' file
}
\description{
Function used to create the GUI for selecting the MODIS images download and preprocessing options
}
\details{
Allows to select the folder where MODIS original and preprocessed files will be saved, the starting and ending processing years and 
the reprocessing and redownloading options. Also calls the MOD_Proc routine if the user selects "Start".
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Oct 26, 2012
}

