% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_moddownload.R
\name{frg_moddownload}
\alias{frg_moddownload}
\title{frg_moddownload}
\usage{
frg_moddownload(OutOrig_Path, ReProc, yy)
}
\arguments{
\item{OutOrig_Path}{string Main folder where the original MODIS mosaics
(i.e., one pan european image for each data and data type) will be stored}

\item{ReProc}{Flag. If = 1, existing preporocessed images will be recomputed
and overwritten}

\item{yy}{Year for which images are to be downloaded and processed}
}
\description{
Function used to download, mosaic and reproject MODIS data
}
\details{
The function leverages
functionality form \code{MODIStsp} package \link{https://github.com/lbusett/MODIStsp/}
\strong{Download and Mosaicing}:
Images covering Europe for DOYS 209 and 225 of each selected year are downloaded,
mosaiced and reprojected in LAEA projection, with 250m pixels size
Resulting images are saved in TIFF format in the Originals subfolder of the
output folder selected by the user to store MODIS imafes. Separate subfolders are used
to store images related to  NDVI, QA (Quality Assurance) and Pixel Reliability.
Original HDF files are stored in the "hdf" folder, while METADATA files to
access the full time series in \code{ENVI} and \code{R} are stored in the \code{time_series}
subfolder
}
\author{
Lorenzo Busetto, PhD (2017) - email lbusett@gmail.com
}
