% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_plotstat_multiple.R
\name{frg_plotstat_multiple}
\alias{frg_plotstat_multiple}
\title{frg_plotstat_multiple}
\usage{
frg_plotstat_multiple(in_file, out_file, opts)
}
\arguments{
\item{in_file}{string File of scaled VI Time Series extracted for multiple-fire
areas}

\item{out_file}{string Basename for output files}

\item{opts}{\code{list} of options passed from \code{frg_fullprocessing()}}
}
\value{
Saves a .RData file containing data needed to plot Med_SVI time series
for the different multiple-fire BAs
}
\description{
Function used extract the info necessary to plot the Med_SVI
time series (i.e., boxplots, overlapid of BA, fireyears, ecc) for areas
burnt multiple times
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
