% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_process_modis.R
\name{frg_process_modis}
\alias{frg_process_modis}
\title{frg_modproc}
\usage{
frg_process_modis(opts, UI_check, max_UI, force_update)
}
\arguments{
\item{UI_check}{flag If = 1 then a chack on pixel usefulness index is done while
determining pixels to be used to compute average NDVI. default = 1}

\item{max_UI}{Max UI value kept when averaging MODIS data. default = 5}

\item{mod_dir}{string Folder where the original and preprocessed image will be stored}

\item{opts$start_year}{numeric Starting year of the analysis}

\item{opts$end_year}{numeric Ending Year of the analysis}

\item{opts$reproc_images}{numeric if = 1, already existing MODIS mosaics will be reprocessed}

\item{opts$redown}{numeric if = 1, MODIS images needed to create already existing mosaic
files will be opts$redownloaded, and already existing mosaics will be overwritten}
}
\description{
Function used to download, mosaic and preprocess MODIS data
}
\details{
Function used to download, mosaic and preprocess MODIS data.
\enumerate{
\item \strong{Download and Mosaicing}:
Images covering Europe for end July - mid august  of each selected
year are downloaded, mosaiced and reprojected
Resulting images are saved in TIFF format in the Originals subfolder
of the output folder selected by the user. Separate subfolders are used
to store images related to  NDVI, RED and NIR reflectance, QA
(Quality Assurance) and Pixel Reliability.
\item \strong{Preprocessing}:
The preprocessing steps conducted are the following:
a. MODIS UI (Usefulness Index) values are extracted from the MODIS QA
(Quality Assurance) files
c. The average yearly values for DVI and NDVI are computed starting from the
images of DOY 209 and 225. NODATA and bad quality data#'   are automatically removed
}
}
\author{
Lorenzo Busetto (2016)
email: lbusett@gmail.com
}
