% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_Process_Shapefile.R
\name{FRG_Process_Shapefile}
\alias{FRG_Process_Shapefile}
\title{FRG_Process_Shapefile}
\usage{
FRG_Process_Shapefile(Shape_File = Shape_File,
  Intermediate_Folder = Intermediate_Folder)
}
\arguments{
\item{Shape_File}{string Input Shapefile of BAs}

\item{Intermediate_Folder}{string Folder where the new shapefiles will be placed}

\item{CLC_File_00}{string ENVI file containing the CORINE land Cover map 2000, recoded to the EFFIS legend}
}
\description{
Function used Process the Burnt area shapefile and create the shapefile of area burnt once \cr
             and that of areas burnt multiple times and the corresponding ROIs
}

