% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_recov_statmatrix.R
\name{frg_recov_statmatrix}
\alias{frg_recov_statmatrix}
\title{frg_recov_statmatrix}
\usage{
frg_recov_statmatrix(plot_stat, Data_Median, MedWdt)
}
\arguments{
\item{plot_stat}{data passed from frg_sigmatrix}

\item{Data_Median}{data passed from frg_sigmatrix}

\item{MedWdt}{data passed from frg_sigmatrix}
}
\value{
Data frame containing info on Fire characteristics (ID, Year,
CLC_Class, ecc + Info on results of significance analysis)
}
\description{
Helper function used to add additional info to the signiicance
matrix computed in \code{frg_sigmatrix}
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
