% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_sigyears.R
\name{FRG_RecovStat_Matrix}
\alias{FRG_RecovStat_Matrix}
\title{FRG_RecovStat_Matrix}
\usage{
FRG_RecovStat_Matrix(file_in = file_in, plot_stat = plot_stat,
  Data_Median = Data_Median, MedWdt = MedWdt)
}
\arguments{
\item{file_in}{}

\item{plot_stat}{}

\item{Data_Median}{}

\item{MedWdt}{}
}
\description{
FRG_RecovStat_Matrix
}
