% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_helpers.R
\name{frg_set_arcpypath}
\alias{frg_set_arcpypath}
\title{frg_set_arcpypath}
\usage{
frg_set_arcpypath()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
check if arcpypath global option was already set.
Otherwise ask to set it and check it.
}
\details{
DETAILS
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
