% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_helpers.R
\name{frg_set_effispath}
\alias{frg_set_effispath}
\title{FUNCTION_TITLE}
\usage{
frg_set_effispath()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
check if effispath global option was already set.
Otherwise ask to set it and check it.
}
\details{
DETAILS
}
\examples{
\dontrun{
#EXAMPLE1
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
