% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_SVI_GUI.R
\name{FRG_SVI_GUI}
\alias{FRG_SVI_GUI}
\title{FRG_SVI_Sel}
\usage{
FRG_SVI_GUI()
}
\value{
If the user chooses 'Start' and all processing parameters are OK, the selected parameters are saved in the 'FRG_SVI_Previous.RData' file and the computation starts \cr
Results are saved in appropriate subfolders of the MOD_Dir folder.
}
\description{
Function used to create the GUI for selecting files and options for computation of Scaled Indexes images from MODIS yearly data
}
\details{
Allows to select:
\itemize{
\item MOD_Dir The Main Folder containing the preprocessed MODIS images. This is the folder selected by the user in the Downloading/preprocessing phase (e.g., MODIS_Data)
\item CLC_File_00 File corresponding to the recoded CORINE LAND COVER 2000 map (ENVI format !)
\item Shape_File Shapefile of burnt areas to be considered in the analysis
\item Start_Year Starting year for the analysis
\item End_Year End year for the analysis
\item NKer Width in Km of the moving window considered for computation of scaled indexes
\item Method Method used for Scaled indexes computation. Options are: \cr
		  Forced to "2" = 'Percentage Deviation to Kernel Median': For each pixel, the algorithm identifies all pixels of the image corresponding to the same CLC class contained in a window of width NKer
			It then computes the median of the distribution of values of the analyzed Vegetation Index in the selecteds pixels. Scaled Index for the pixel are then computed as: 
			sVI = 100 * (VI_pix - VI_median)/(VI_median). Previous methods no longer used !
\item SNDVI, SRDVI checkboxes used to select which vegetation indexes are to be used in the analysis
\item ReProc Checkbox used to specify if already present images are to be reprocessed. Defaults to FALSE
}
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Oct 26, 2012
}

