% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_update_oracle.R
\name{frg_update_oracle}
\alias{frg_update_oracle}
\title{frg_update_oracle}
\usage{
frg_update_oracle()
}
\value{
The function is called for its side effects
}
\description{
Function used to update the Oracle database tables using
processing results of FRG
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
