% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_wilcox.R
\name{frg_wilcox}
\alias{frg_wilcox}
\title{FRG_Comp_Sig_Matrix}
\usage{
frg_wilcox(in_file, out_file_pmat, opts)
}
\arguments{
\item{in_file}{string File of scaled VI Time Series extracted for single-fire areas}

\item{out_file_pmat}{\code{character} path to the file to be used to store the
p-values matrix}

\item{opts}{\code{list} of options passed from \code{frg_fullprocessing()}}
}
\value{
Saves a p_matrix.RData file containing the p-values matrix and the
'plot_stats' data frame in the 'p_matrixes' subfolder
}
\description{
Function used to perform the Wilcoxon test of significant differences,
using differnt number of 'before fire years' as reference. The function creates
and saves matrixes of p-values obtained using differnt number of before fire
years. Matrix is saved within the 'p-matrixes' subfolder of the folder
containing the final output file. Additional info necessary to plot the time
series (i.e., boxplots, objectids of fires, fireyears, ecc) are also saved
in the same file (plot_stat data frame).
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
