% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_Full_Processing.R
\docType{data}
\name{Shape_File}
\alias{Shape_File}
\title{FRG_Full_Processing}
\format{An object of class \code{character} of length 1.}
\usage{
Shape_File
}
\arguments{
\item{MOD_Dir}{string Folder where the original and preprocessed image will be stored (i.e., the one above the "Originals" folder)}

\item{Shape_File}{string Input Shapefile of BAs}

\item{CLC_File_00}{string ENVI file containing the CORINE land Cover map 2000, recoded to the EFFIS legend}

\item{Out_Folder}{string Main Results Folder where results of the analysis will be saved}

\item{Start_Year}{numeric Starting year of the analysis}

\item{End_Year}{numeric Ending Year of the analysis}

\item{Method}{numeric. 2 = Percentage Median difference computation \cr}

\item{SRDVI}{numeric. 1 = compute SRDVI; 2 =  Don't Compute SRDVI}

\item{SNDVI}{numeric. 1 =  compute SNDVI; 2 = Don't Compute SNDVI}

\item{ReProc}{numeric if = 1, already existing Scaled Indexes will be recomputed}

\item{ReDown}{numeric if = 1, MODIS images needed to create already existing mosaic files will be redownloaded, and 
already existing mosaics will be overwritten}

\item{ReProcIm}{numeric if = 1, already existing MODIS mosaics will be reprocessed}

\item{erode}{flag if = 1, analysis is conducted only on core pixels (defaults to 1)}

\item{min_pix}{numeric minimum number of core pixels needed for a BA to be processed (default to 10)}

\item{NKer}{numeric width (in Km) of the moving window used for computation of scaled indexes}

\item{sig_level}{numeric Significance level for Wilcoxon test. Default to 0.05}

\item{sub_zones}{Obsolete}

\item{perc_diffs}{numeric hash table Reduction Threshold on NDVIR reduction used in significance reduction analysis with wilcoxon test}
}
\description{
Function used to apply all FRG processing steps \cr
}
\author{
Lorenzo Busetto - email: lorenzo.busetto@jrc.ec.europa.eu
}
\keyword{datasets}

