% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_buildroi.R
\name{frg_buildroi}
\alias{frg_buildroi}
\title{frg_buildroi}
\usage{
frg_buildroi(Shape_File, CLC_File_00, exp_path_str, ROI_file)
}
\arguments{
\item{Shape_File}{string Input Shapefile of BAs}

\item{CLC_File_00}{string ENVI file containing the CORINE land Cover map 2000, recoded to the EFFIS legend}
}
\description{
Accessory function to call IDL routine for ROI creation.
}
\details{
Function used to create a batch file (\code{FRG_Create_ROI_batch.pro}) which
is then used to call the \code{FRG_CREATE_ROI.pro} IDL unction from a command shell.

\code{FRG_CREATE_ROI.pro} is used to create a ROI file starting from the shapefile
of burned areas. See \code{FRG_CREATE_ROI.pro} in \code{/IDL/Build_ROIS} for further
documentation
}
