% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_compmean.R
\name{frg_compmean}
\alias{frg_compmean}
\title{frg_compmean}
\usage{
frg_compmean(OutOrig_Path, yy, UI_check, max_UI, force_update)
}
\arguments{
\item{OutOrig_Path}{string Main folder where the original MODIS mosaics
(i.e., one pan european image for each data and data type) will be stored}

\item{yy}{Year for which images are to be downloaded and processed}

\item{UI_check}{}

\item{max_UI}{}

\item{ReProc}{}
}
\description{
frg_compmean
}
\author{
Lorenzo Busetto, PhD (2012)
email: lbusett@gmail.com
}
