% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_compmean.R
\name{frg_compmean}
\alias{frg_compmean}
\title{frg_compmean}
\usage{
frg_compmean(opts, yy, UI_check, max_UI, force_update)
}
\arguments{
\item{UI_check}{}

\item{max_UI}{}

\item{ReProc}{}
}
\description{
frg_compmean
}
\author{
Lorenzo Busetto, PhD (2012)
email: lbusett@gmail.com
}
