% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_config.R
\name{frg_config}
\alias{frg_config}
\title{FUNCTION_TITLE}
\usage{
frg_config(cfg_dir)
}
\arguments{
\item{cfg_dir}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
