% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_createmask.R
\name{frg_createmask}
\alias{frg_createmask}
\title{frg_createmask}
\usage{
frg_createmask(roi_file, CLC_File_00, exp_path_str, FireMask_File, force_update)
}
\arguments{
\item{roi_file}{}

\item{CLC_File_00}{string ENVI file containing the CORINE land Cover map 2000, recoded to the EFFIS legend}
}
\description{
Accessory function to call IDL routine for Mask creation.
}
\details{
Function used to create a batch file (\code{FRG_CreateMask_Batch.pro}) which
is then used to call the \code{FRG_Create_Mask.pro} IDL function from a command shell.

\code{FRG_Create_Mask.pro} is used to create a binary mask to exclude burned areas
from statistics computaitons. See \code{FRG_Create_Mask.pro} in \code{/IDL/VI_Elaborations}
for further documentation
}
