% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_createmaskeroded.R
\name{frg_createmask_eroded}
\alias{frg_createmask_eroded}
\title{frg_createmask}
\usage{
frg_createmask_eroded(opts, exp_path_str, force_update)
}
\arguments{
\item{opts$roi_file}{}
}
\description{
Accessory function to call IDL routine for Mask creation.
}
\details{
Function used to create a batch file (\code{FRG_CreateMask_Eroded_Batch.pro}) which
is then used to call the \code{FRG_Create_Mask_Eroded.pro} IDL function from a command shell.

\code{FRG_Create_Mask_Eroded.pro} is used to create an eroded version of the binary mask
of burned areas, which allows to identify non-border (CORE) burnt pixels.
See \code{FRG_Create_Mask_Eroded.pro} in \code{/IDL/VI_Elaborations} for further documentation
}
