% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_createmeta.R
\name{frg_createmeta}
\alias{frg_createmeta}
\title{frg_createmeta}
\usage{
frg_createmeta(Index, Start_Year, End_Year, Method, Out_Folder, Out_Files)
}
\arguments{
\item{Start_Year}{numeric Starting year of the analysis}

\item{End_Year}{numeric Ending Year of the analysis}

\item{Method}{numeric. 2 = Percentage Median difference computation}

\item{Out_Folder}{string Main Results Folder where results of the analysis will be saved}
}
\description{
frg_createmeta
}
