% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_def_files.R
\name{frg_def_files}
\alias{frg_def_files}
\title{frg_def_files}
\usage{
frg_def_files(opts)
}
\arguments{
\item{opts}{initial options list created by frg_main}
}
\value{
\code{opts} list containing all main file names and folders
}
\description{
Helper function used to define required file names and folders
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
