% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_def_files.R
\name{frg_def_files}
\alias{frg_def_files}
\title{frg_def_files}
\usage{
frg_def_files(Out_Dir, Start_Year, End_Year, MOD_Dir, Shape_File)
}
\arguments{
\item{Out_Dir}{}

\item{Start_Year}{}

\item{End_Year}{}

\item{MOD_Dir}{}

\item{Shape_File}{}
}
\value{
ff list contains all main file names and folders
}
