% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_fastzonal.R
\name{frg_fastzonal}
\alias{frg_fastzonal}
\title{frg_fastzonal}
\usage{
frg_fastzonal(in_rts, zone_object, clc_object, mask_object, start_date = NULL,
  end_date = NULL, start_band = NULL, end_band = NULL, id_field = NULL,
  FUN = "null", out_format = "xts", small = TRUE,
  small_method = "centroids", na.rm = TRUE, verbose = FALSE,
  maxchunk = 5e+07)
}
\arguments{
\item{in_rts}{a}

\item{zone_object}{f}

\item{mask_object}{g}

\item{end_date}{g}

\item{start_band}{t}

\item{end_band}{re}

\item{id_field}{g}

\item{FUN}{h}

\item{out_format}{g}

\item{small}{h}

\item{small_method}{h}

\item{na.rm}{g}

\item{verbose}{g}

\item{maxchunk}{f}
}
\value{
ts_data data.table containing SVI time series extracted for each burnt
area from \code{in_rts} multitemporal SVI raster file
}
\description{
aaaa
}
\examples{

}
