% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_fastzonal.R
\name{frg_fastzonal}
\alias{frg_fastzonal}
\title{Helper function used to extract data from the Med_SNDVI time series
in correspondance of the different burnt areas}
\usage{
frg_fastzonal(in_rts, zone_object, clc_object, mask_object, start_date = NULL,
  end_date = NULL, start_band = NULL, end_band = NULL, id_field = NULL,
  FUN = "null", out_format = "xts", small = TRUE,
  small_method = "centroids", na.rm = TRUE, verbose = FALSE,
  maxchunk = 5e+07)
}
\arguments{
\item{in_rts}{PARAM_DESCRIPTION}

\item{zone_object}{PARAM_DESCRIPTION}

\item{clc_object}{PARAM_DESCRIPTION}

\item{mask_object}{PARAM_DESCRIPTION}

\item{start_date}{PARAM_DESCRIPTION, Default: NULL}

\item{end_date}{PARAM_DESCRIPTION, Default: NULL}

\item{start_band}{PARAM_DESCRIPTION, Default: NULL}

\item{end_band}{PARAM_DESCRIPTION, Default: NULL}

\item{id_field}{PARAM_DESCRIPTION, Default: NULL}

\item{FUN}{PARAM_DESCRIPTION, Default: 'null'}

\item{out_format}{PARAM_DESCRIPTION, Default: 'xts'}

\item{small}{PARAM_DESCRIPTION, Default: TRUE}

\item{small_method}{PARAM_DESCRIPTION, Default: 'centroids'}

\item{na.rm}{PARAM_DESCRIPTION, Default: TRUE}

\item{verbose}{PARAM_DESCRIPTION, Default: FALSE}

\item{maxchunk}{PARAM_DESCRIPTION, Default: 5e+07}
}
\value{
ts_data data.table containing SVI time series extracted for each burnt
area from \code{in_rts} multitemporal SVI raster file
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
