% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_fullproc_gui.R
\name{frg_fullproc_gui}
\alias{frg_fullproc_gui}
\title{frg_fullproc_gui}
\usage{
frg_fullproc_gui()
}
\description{
Function used to create the GUI for selecting the parameters for
full FRG procesing
}
\details{
Main GUI of the tool - used only to set the processing options and
call the appropraite routines for updating results (FRG_Full_Processing.R)
If the user chooses 'Start' and all processing parameters are OK, the selected
parameters are saved in the 'FRG_Full_Previous.RData' file and processing is
started
}
\author{
Lorenzo Busetto (2012 - 2017)
email: lbusett@gmail.com
}
