% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_fullprocessing.R
\name{frg_fullprocessing}
\alias{frg_fullprocessing}
\title{FUNCTION_TITLE}
\usage{
frg_fullprocessing(opts, force_update, MOD_dwl = FALSE, Comp_SVI = TRUE,
  Extr_Stat = TRUE, Sig_Anal = TRUE)
}
\arguments{
\item{opts}{PARAM_DESCRIPTION}

\item{force_update}{PARAM_DESCRIPTION}

\item{MOD_dwl}{PARAM_DESCRIPTION, Default: FALSE}

\item{Comp_SVI}{PARAM_DESCRIPTION, Default: TRUE}

\item{Extr_Stat}{PARAM_DESCRIPTION, Default: TRUE}

\item{Sig_Anal}{PARAM_DESCRIPTION, Default: TRUE}
}
\description{
FUNCTION_DESCRIPTION
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
