% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_MOD_Download.R
\name{FRG_MOD_Download}
\alias{FRG_MOD_Download}
\title{FRG_MOD_Proc}
\usage{
FRG_MOD_Download(dates = dates, OutOrig_Path = OutOrig_Path,
  ReDown = ReDown, yy = yy)
}
\arguments{
\item{dates}{string array of dates to be downloaded}

\item{OutOrig_Path}{string Main folder where the original MODIS mosaics (i.e., one pan european image for each data and data type) will be stored}

\item{ReDown}{numeric if = 1, MODIS images needed to create already existing mosaic files will be redownloaded, and
already existing mosaics will be overwritten}

\item{yy}{numeric Year considered}
}
\value{

}
\description{
Function used to download, mosaic and reproject MODIS data
}
\author{
Babak Naimi (naimi@r-gis.net)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Oct 29, 2012
}

