% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_MOD_Proc.R
\name{FRG_MOD_Proc}
\alias{FRG_MOD_Proc}
\title{FRG_MOD_Proc}
\usage{
FRG_MOD_Proc(MOD_Dir = MOD_Dir, Start_Year = Start_Year,
  End_Year = End_Year, ReProcIm = ReProcIm, ReDown = ReDown,
  UI_check = 1, max_UI = 5)
}
\arguments{
\item{MOD_Dir}{string Folder where the original and preprocessed image will be stored}

\item{Start_Year}{numeric Starting year of the analysis}

\item{End_Year}{numeric Ending Year of the analysis}

\item{ReProcIm}{numeric if = 1, already existing MODIS mosaics will be reprocessed}

\item{ReDown}{numeric if = 1, MODIS images needed to create already existing mosaic files will be redownloaded, and
already existing mosaics will be overwritten}

\item{max_UI}{Max UI value kept when averaging MODIS data. default = 5}

\item{UI_Check}{flag If 1 , then check on usefulness index is performed when averaging MODIS data. default = 1}
}
\value{

}
\description{
Function used to download, mosaic and preprocess MODIS data
}
\details{
Function used to download, mosaic and preprocess MODIS data. \cr
	\itemize{
		\item Download and Mosaicing: \cr
              Images covering Europe for end July - mid august  of each selected year are downloaded, mosaiced and reprojected \cr
              Resulting images are saved in TIFF format in the Originals subfolder of the output folder selected by the user. Separate subfolders are used\cr
              to store images related to  NDVI, RED and NIR reflectance, QA (Quality Assurance) and Pixel Reliability.
	\item Preprocessing: \cr
              The preprocessing steps conducted are the following: \cr
              1) MODIS UI (Usefulness Index) values are extracted from the MODIS QA (Quality Assurance) files
              2) The DVI (Difference Vegetation Index) is computed for each date starting from the RED and NIR MODIS images
              3) The average yearly values for DVI and NDVI are computed starting from the images of DOY 209 and 225. NODATA and bad quality data\cr
                  are automatically removed
}
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Oct 29, 2012
}

