% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_process_shapefile.R
\name{frg_process_shapefile}
\alias{frg_process_shapefile}
\title{frg_process_shapefile}
\usage{
frg_process_shapefile(Shape_File, Intermediate_Folder)
}
\arguments{
\item{Shape_File}{PARAM_DESCRIPTION}

\item{Intermediate_Folder}{PARAM_DESCRIPTION}
}
\value{
The function is called for its side effects
}
\description{
Process the Burnt area shapefile to create the shapefiles of areas
burnt once and that of areas burnt multiple times and the corresponding CSV
LUT by calling the "FRG_Burnt_Areas_Processing.py" python script
}
\author{
Lorenzo Busetto, phD \href{mailto:lbusett@gmail.com}{lbusett@gmail.com},
Giovanni Caudullo
}
