% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_helpers.R
\name{frg_set_arcpypath}
\alias{frg_set_arcpypath}
\title{FUNCTION_TITLE}
\usage{
frg_set_arcpypath()
}
\value{
OUTPUT_DESCRIPTION

path to argis python
}
\description{
check if effispath global option was already set.
Otherwise ask to set it and check it.

check if arcpypath global option was already set.
Otherwise ask to set it and check it.
}
\details{
DETAILS
}
\examples{
\dontrun{
#EXAMPLE1
 }
}
\author{
Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}

Lorenzo Busetto, phD (2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
