% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_sigmatrix.R
\name{frg_sigmatrix}
\alias{frg_sigmatrix}
\title{frg_sigmatrix}
\usage{
frg_sigmatrix(in_file, out_file, opts)
}
\arguments{
\item{in_file}{string File of scaled VI Time Series extracted for single-fire areas}

\item{out_file}{string Basename for output files}

\item{opts}{\code{list} of options passed from \code{frg_fullprocessing()}}
}
\value{
Saves a .RData file containing data needed to plot NDVIR time series for the different single-fire BAs
}
\description{
Function used to perform statistical analysis in order to
determine significance of scaled VI reductions in years following the fire.
\itemize{
\item Starting from extracted time series, compute matrixes of p-values using
FRG_Comp_Regr_Matrix. p-values matrixes contain result of the Wilcoxon
test, computed using the selected number of reference pre-fire years and
the selected minimum difference percentage.
\item Matrixes of significances are then created using FRG_Significance_Analysis and saved.
}

Note: this is only a wrapper used to call the FRG_Comp_Regr_Matrix
and FRG_Significance_Analysis functions)
}
\author{
Lorenzo Busetto, phD (2013-2017) \href{mailto:lbusett@gmail.com}{lbusett@gmail.com}
}
