% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_SVI_GUI.R
\name{FRG_SVI_GUI}
\alias{FRG_SVI_GUI}
\title{FRG_SVI_Sel}
\usage{
FRG_SVI_GUI()
}
\value{
If the user chooses 'Start' and all processing parameters are OK, the selected parameters are saved in the 'FRG_SVI_Previous.RData' file and the computation starts \cr
Results are saved in appropriate subfolders of the MOD_Dir folder.
}
\description{
Function used to create the GUI for selecting files and options for computation of Scaled Indexes images from MODIS yearly data
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Oct 26, 2012
}

