% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frg_update_oracle.R
\name{frg_update_oracle}
\alias{frg_update_oracle}
\title{frg_update_oracle}
\usage{
frg_update_oracle(effis_folder)
}
\value{
In_folder : Name of the folder containing the files used to update the oracle tables
}
\description{
Function used to update the Oracle database tables using processing results of FRG
}
\details{
This function is used to update the Oracle database tables using processing results of FRG
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Nov 8, 2012
}
