% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FRG_Update_Oracle.R
\name{FRG_Update_Oracle}
\alias{FRG_Update_Oracle}
\title{FRG_Update_Oracle}
\usage{
FRG_Update_Oracle(effis_folder)
}
\value{
In_folder : Name of the folder containing the files used to update the oracle tables
}
\description{
Function used to update the Oracle database tables using processing results of FRG
}
\details{
This function is used to update the Oracle database tables using processing results of FRG
}
\author{
Lorenzo Busetto (2012)
email: lorenzo.busetto@jrc.ec.europa.eu

Created Date: Nov 8, 2012
}
